progress : very near beta
time limit for E3D and dwindow.

git-svn-id: svn://127.0.0.1@97 2ece6d47-a862-c445-b01b-7049066ce7d0

+++ dwindow/DShowSubtitleRenderer.cpp
@@ -36,7 +36,7 @@ HRESULT DShowSubtitleRenderer::SampleCB(IMediaSample *sample)
 
 	//CAutoLock lck(&m_subtitle_sec);
 	if (m_srenderer)
-		m_srenderer->add_data(p, sample->GetActualDataLength(), (start+m_subtitle_seg_time)/10000, (end+m_subtitle_seg_time)/10000);
+		m_srenderer->add_data(p, sample->GetActualDataLength(), (int)((start+m_subtitle_seg_time)/10000), (int)((end+m_subtitle_seg_time)/10000));
 
 	return S_OK;
 }
